#37 - refactor filtering
